requirements.txt
# CryptoV4 Dependencies
# API and Web Framework
flask>=2.0.2
flask-cors>=3.0.10
gunicorn>=20.1.0
python-dotenv>=0.19.2

# Data Processing
pandas>=1.3.5; python_version < '3.12'
pandas>=1.5.0; python_version >= '3.12'
numpy>=1.20.0; python_version < '3.12'
numpy>=1.22.0; python_version >= '3.12'
matplotlib>=3.5.1
scikit-learn>=1.0.2
pytz>=2021.3

# Crypto APIs
python-binance>=1.0.15
newsapi-python>=0.2.6
cryptocompare>=0.7.5

# Database
pymongo>=4.0.1
dnspython>=2.2.0

# Sentiment Analysis and NLP
# Optional packages - install as needed:
# transformers>=4.16.2; python_version < '3.12'
# transformers>=4.30.0; python_version >= '3.12'
nltk>=3.6.7
# For newer Python versions, specify platform-specific torch installation
# pip install torch==1.10.2+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
vaderSentiment>=3.3.2

# Technical Analysis
ta>=0.9.0
# TA-Lib is optional due to C dependencies
# ta-lib>=0.4.24

# Utilities
requests>=2.27.1
python-dateutil>=2.8.2
tqdm>=4.62.3

# Testing
pytest>=7.0.0
pytest-mock>=3.7.0
mock>=4.0.3

# Development
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1

# Add ccxt for exchange interface
ccxt>=1.60.0 